Android. Making Ogre compile with standalone toolchain
Issue is with included header files. For example: `#include <GLES3/gl3.h>`. We have such file in Ogre sources: `RenderSystems/GLES2/include/GLES3/gl3.h`, and we want to use it. But same file exists in system headers, so, during compilation file from system headers is used. It does not have some tipes and compilation fails.
We could fix that by playing with order of header search paths. But pretty complicated to dig in into Ogre3D CMake scripts and make changes there.
As a solution, those files were moved to ewn folder 'Bundled'. So, we explicitly are saying, which files we want to include.