Fix EGL on Linux
Besides adding a few casts to make it build, we also make some changes
to make it actually work. The key moment was that Ogre thought they could
call eglCreateWindowSurface() without specifying the display and condig,
and then get both from the surface. That may have worked in the past,
but it no longer does.

This fix may or may not work when the "externalWindowHandle"
configuration option is not used. On Linux, Fatmap always uses it.