Fix uninitialized member mVisible in Ogre::GLXWindow class
The symptoms of this problem is that sometimes on Linux we
get only a blank window.

The fix is to set mVisible to false in the constructor and
to true in Ogre::GLXWindow::create(), as done in
Ogre::X11EGLWindow, except unlike us, it doesn't bother to
initialize it in the constructor.