Implement a fallback for Ogre::Font::getGlyphInfo()
Issue:
When the caption in a ParamsLabel has characters missing from a font,
an exception is thrown from Ogre::Root::renderOneFrame().

Solution:
Make Ogre::Font::getGlyphInfo() return the information for the
fallback '?' character when the requested character is missing
from the font. Since the glyph information also contains the
UV coordinates of the character, this change is enough to display
question marks in place of characters missing from the font.