Add support for ANGLE on iOS
A new switch "OGRE_GLES2_USE_ANGLE" was introduced to switch between
EAGL and ANGLE on iOS. Other than that, we choose to implement ANGLE
support in such a way to require minimal (if any) modifiactions on
the client side. In particular, the EAGL2View and EAGL2ViewController
classes are still available, even though they no longer use the EAGL
API. However, EAGL2Window is not available in ANGLE mode. In fact, we
chose not to support Ogre creating a window by itself in ANGLE mode.
That was done to simplify the implementation. Therefore, the client
is expected to pass either the "externalViewHandle" option or the
"externalViewControllerHandle" one with view already set on it.

Besides setting OGRE_GLES2_USE_ANGLE to TRUE, you have to set
ANGLE_INCLUDE_DIR and ANGLE_FRAMEWORKS_DIR to point to appropriate
locations. Ogre will use the libGLESv2.framework and libEGL.framework
located in ANGLE_FRAMEWORKS_DIR.