Unset the current GL context in _notifySurfaceDestroyed()
In order not to lose the GL context when the app goes into background,
Ogre seems to be using a trick suggested here [1], except it fails
to unset the current context before destroying the surface. This commit
does precisely that.

Now, most likely the GL context may still be lost in low resource
situations [2], so this change is unlikely to help, but it's still
worth a try.

[1]: https://groups.google.com/g/android-ndk/c/jwVMF6zINus
[2]: https://github.com/godotengine/godot/issues/22955